Use a dhcp exit hook to restore rt1
Without this patch the two entries in rt1 are lost when dhclient
reconfigures eth1.  This situation arises in at least two places:

 1. Race condition during startup, google-network-daemon executes
    concurrently with startup-script runner
 2. `systemctl restart google-network-daemon`.

This patch fixes the problem by configuring
`/etc/dhcp/dhclient-exit-hooks.d/vpc-link.sh` with the behavior of
restoring rt1 when eth1 is plumbed.

Note this is only tested for eth1, not for eth2 ... eth7.