Manage multinic instance groups in all zones in the region
Without this patch the `52_regional_multinic` module makes the blind
assumption zones a, b, and c exist in the region.  This is a problem for
us-east1 and europe-west1 which do not have zone a.

This patch fixes the problem by changing 50_compute's zone input
parameter to a list of zone name strings, `zones`.  Similarly,
`52_regional_multinic` is updated to automatically determine all
available zones in the region and create multinic groups in each zone.

Resolves: https://github.com/openinfrastructure/terraform-google-multinic/issues/20