Document zero-downtime major version upgrade process
This patch documents a process to upgrade to new major versions of this
module without incurring any downtime.  The module changes to support
this by relying on name_prefix and route priority to temporarily stand
up two side-by-side copies of the multinic infrastructure, cut over to
the new version, then delete the previous version.

Resolves: https://github.com/openinfrastructure/terraform-google-multinic/issues/23