Fix Error: Invalid prefix for given prefix length
This patch fixes the error caused by using the IP address instead of the
network address when replacing routes.  The error doesn't affect the
behavior of the system, but it is confusing and a red-herring when
troubleshooting other issues.  The `set -x` is intentionally left in to
help troubleshoot any future issues, the commands are displayed in the
console and journal in-line with the commands generating the output.

See: [Issue 28][1] for more details and debug logs.

[1]: https://github.com/openinfrastructure/terraform-google-multinic/issues/28