Monitor the health of IP forwarding
If the rules or routing table change after the policy-routing.service
starts, then stop the hc-health.service to trigger auto-healing.

The instance is marked as unhealthy in the following conditions.  Note,
viaeth0 and viaeth1 tables and rules are the focus because the Google
Guest OS agent dynamically modifies the local routing table in response
to forwarding rules being added or removed.

 * The routes in table viaeth0 change.
 * The routes in table viaeth1 change.
 * Any routing rule changes, which could affect if viaeth0 or viaeth1
   tables are looked up.
 * IP Forwarding is not enabled in the kernel.